Fix configuration edge cases (#453)
* Fix crash when not using jailer

* Improve system() wrapper for debug

Now in case a command failed it tell us which function called it

* Problem: command warnings if no jailman user exist

even with USE_JAILER=False

Solution: This was caused by prepare_jailer() even thought it was not needed.
Made it's running conditional on use_jailer

* Problem: Crash when lack of perm for nftables

e.g: When running as not root and with network disable
this was caused by the json.loads() unable to parse the empty output

Solution: Return a default value

* Problem: Crash on VM clean up if no tap interface

Solution: made conditional

* Problem: Status endpoint was hanging for unhandled error

it was stuck and never returned anything

Solution: Had a catchall for the exception that were not handled before

* Problem: ndppd error even if USE_NDP_PROXY=0

Solution: Make check conditional

* fix isort

* Update src/aleph/vm/orchestrator/run.py


---------

+++ src/aleph/vm/controllers/firecracker/executable.py
@@ -21,7 +21,7 @@ from aleph.vm.guest_api.__main__ import run_guest_api
 from aleph.vm.hypervisors.firecracker.microvm import FirecrackerConfig, MicroVM
 from aleph.vm.network.firewall import teardown_nftables_for_vm
 from aleph.vm.network.interfaces import TapInterface
-from aleph.vm.storage import get_volume_path
+from aleph.vm.storage import chown_to_jailman, get_volume_path
 
 try:
     import psutil  # type: ignore [no-redef]
@@ -250,9 +250,11 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType]):
             logger.debug("setup done")
         except Exception:
             # Stop the VM and clear network interfaces in case any error prevented the start of the virtual machine.
+            logger.error("VM startup failed, cleaning up network")
             await self.fvm.teardown()
             teardown_nftables_for_vm(self.vm_id)
-            await self.tap_interface.delete()
+            if self.tap_interface:
+                await self.tap_interface.delete()
             raise
 
         if self.enable_console:
@@ -280,7 +282,7 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType]):
         self.guest_api_process.start()
         while not exists(vsock_path):
             await asyncio.sleep(0.01)
-        subprocess.run(f"chown jailman:jailman {vsock_path}", shell=True, check=True)
+        await chown_to_jailman(Path(vsock_path))
         logger.debug(f"started guest API for {self.vm_id}")
 
     async def stop_guest_api(self):
@@ -291,7 +293,8 @@ class AlephFirecrackerExecutable(Generic[ConfigurationType]):
         if self.fvm:
             await self.fvm.teardown()
             teardown_nftables_for_vm(self.vm_id)
-            await self.tap_interface.delete()
+            if self.tap_interface:
+                await self.tap_interface.delete()
         await self.stop_guest_api()
 
     async def create_snapshot(self) -> CompressedDiskVolumeSnapshot:
